refactor(scripts): share gateway benchmark runtime
